% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdfdr.R
\name{tdfdr.plot}
\alias{tdfdr.plot}
\title{Plot the tdfdr results}
\usage{
tdfdr.plot(tdfdr.obj, fdr.level = seq(0.01, 0.2, len = 20),
  nlimit = 10000, file.name = "tdfdr.plot")
}
\arguments{
\item{tdfdr.obj}{a list returned from running 'tdfdr'.}

\item{fdr.level}{a numeric vector of the target FDR levels to be plotted.}

\item{nlimit}{an integer, the number of insignificant z-scores to be sampled to reduce the visualization complexity.}

\item{file.name}{a character string for the name of the generated pdf file.}
}
\value{
A list of \code{'ggplot2'} objects.
}
\description{
The function plots the number of hits (discoveries) against different FDR levels and plots the unadjusted z-score against adjusted z-score with highlighted discoveries.
}
\examples{
x <- rnorm(50)
z <- x + rnorm(50)
z <- scale(z)
y1 <- x \%*\% t(rep(0.5, 50)) + matrix(rnorm(50 * 50), 50, 50)
y2 <- z \%*\% t(rep(0.5, 50)) + matrix(rnorm(50 * 50), 50, 50)
y3 <- x \%*\% t(rep(0.5, 50)) + z \%*\% t(rep(0.5, 50)) + matrix(rnorm(50 * 50), 50, 50)
y <- cbind(y1, y2, y3, matrix(rnorm(50 * 850), nrow = 50))
obj2 <- tdfdr(y, x, z, alpha = 0.05)
tdfdr.plot(obj2)
}
\references{
Two-dimensional false discovery rate control for powerful confounder adjustment in omics association studies. Submitted.
}
\author{
***
}
\keyword{FDR}
\keyword{confounder}
\keyword{multiple testing}
