% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdfdr.R
\name{tdfdr}
\alias{tdfdr}
\title{Two-dimenisonal false discovery rate control for powerful confounder adjustment in omics-wide association studies}
\usage{
tdfdr(y, x, z, est.pi0 = TRUE, lambda = 0.5, alpha = 0.05,
  etype = c("FDR", "FWER"), t1a = NULL, t2a = NULL, ngrid = 50,
  parallel = FALSE, cores = NULL, verbose = TRUE)
}
\arguments{
\item{y}{a matrix of numeric values for the high-dimensional data (sample size by number of features).}

\item{x}{a vector of numeric values for the variable of interest or a factor of two levels.}

\item{z}{a factor, data.frame, or a matrix of numeric values, representing the confounding variables.}

\item{est.pi0}{a logical value indicating whether John-Storey's correction (pi0 estimation) should be applied. Default TRUE.}

\item{alpha}{a numeric value, the target FDR level.}

\item{etype}{a character string indicating the type of error control ("FDR" or "FWER"). Currently only supports "FDR".}

\item{t1a, t2a}{a numeric vector giving the search grid for the z-score of adjusted and unadjusted statistics, respectively. Default is NULL, and the grid
will be determined automatically.}

\item{ngrid}{an integer number indicating the number of grids. Default is 50. For more refined results, increase the number.}

\item{parallel}{a logical value indicating whether to search in parallel. Default is FALSE.}

\item{cores}{an integer number indicating the number of cores needed if \code{parallel} is TRUE.}

\item{verbose}{a logical value indicating whether to print out progress messages during computation.}
}
\value{
A list with the elements
\item{pos}{a vector of logical values indicating the significant findings (positives) at the given FDR level.}
\item{t1, t2}{a numerical value indicating the best point on the search grid.}
\item{t1t2}{a numeric vector containing the values of the search grid.}
\item{Zu, Za}{a numeric vector of unadjusted or adjusted z-scores.}
\item{NP}{a numeric vector of the number of positives at all the grid points.}
\item{FDP}{a numeric vector of the false discovery proportions at all the grid points.}
\item{p.value}{a numeric vector of the p-values of the adjusted analysis.}
\item{mm}{the object from GLmix, containing the NPEB estimate of the location parameters.}
\item{pi0}{A numeric value of the estimated null proportion.}
}
\description{
The function implements the two-dimensional false discovery rate control for more powerful confounder adjustment in omics-wide association analysis. The method is based on the idea that the confounder(s) usually affect part of the omics features, and thus adjusting the confounder(s)
for ALL omics features will be over-adjustment, leading to reduced statistical power.  The proposed procedure starts with performing
the unadjusted analysis (first dimension - filtering) to narrow down the list of omics features which are more likely to be affected by either the confounder
or the variable of interest or both. In the second dimension, we conduct adjusted analysis on these 'top' candidates to reduce multiple testing burden and increase the power.
In other words, the unadjusted p-values tell us about the probability of the hypotheses being null, and the multiple testing can be focused on those
promising hypotheses. The procedure is theoretically guaranteed to control the false discovery rate while maximizing the power of discovery.
}
\examples{

require(qvalue)
truth <- c(rep(1, 50), rep(0, 50), rep(1, 50), rep(0, 850))
x <- rnorm(50)
z <- x + rnorm(50)
z <- scale(z)

y1 <- x \%*\% t(rep(0.5, 50)) + matrix(rnorm(50 * 50), 50, 50)
y2 <- z \%*\% t(rep(0.5, 50)) + matrix(rnorm(50 * 50), 50, 50)
y3 <- x \%*\% t(rep(0.5, 50)) + z \%*\% t(rep(0.5, 50)) + matrix(rnorm(50 * 50), 50, 50)
y <- cbind(y1, y2, y3, matrix(rnorm(50 * 850), nrow = 50))

# The traditional procedure
obj1 <- summary(lm(y ~ x + z))
pv1 <- sapply(obj1, function(x) x$coefficient[2, 4])
qv1 <- qvalue(pv1)$qvalue
pos1 <- qv1 <= 0.05

# The proposed procedure
obj2 <- tdfdr(y, x, z, alpha = 0.05)
pos2 <- obj2$pos

sum(pos1 & !truth) / max(sum(pos1), 1)
sum(pos2 & !truth) / max(sum(pos2), 1)
sum(pos1 & truth)
sum(pos2 & truth)

}
\references{
Two-dimensional false discovery rate control for powerful confounder adjustment in omics association studies. Submitted
}
\author{
 ***
}
\keyword{FDR}
\keyword{confounder}
\keyword{multiple testing}

