% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdfdr.R
\name{tdfdr.select}
\alias{tdfdr.select}
\title{Select the features at a target FDR level}
\usage{
tdfdr.select(tdfdr.obj, fdr.level = 0.1)
}
\arguments{
\item{tdfdr.obj}{a list returned from running 'tdfdr'.}
\item{fdr.level}{a numeric for the target FDR level.}
}
\value{
A list with the elements
\item{pos}{a vector of logical values indicating the significant findings (positives) at the target FDR level.}
\item{t1, t2}{a numerical value indicating the t1, t2 cutoff values, which achieve the target FDR level.}
}
\description{
The function outputs the features at a given FDR level.
}
\examples{
x <- rnorm(50)
z <- x + rnorm(50)
z <- scale(z)
y1 <- x \%*\% t(rep(0.5, 50)) + matrix(rnorm(50 * 50), 50, 50)
y2 <- z \%*\% t(rep(0.5, 50)) + matrix(rnorm(50 * 50), 50, 50)
y3 <- x \%*\% t(rep(0.5, 50)) + z \%*\% t(rep(0.5, 50)) + matrix(rnorm(50 * 50), 50, 50)
y <- cbind(y1, y2, y3, matrix(rnorm(50 * 850), nrow = 50))
obj2 <- tdfdr(y, x, z, alpha = 0.05)
tdfdr.select(obj2, fdr.level = 0.1)
}
\references{
Two-dimensional false discovery rate control for powerful confounder adjustment in omics association studies. Submitted.
}
\author{
***
}
\keyword{FDR}
\keyword{confounder}
\keyword{multiple testing}
